requirements.txt
python-dotenv==1.0.1
librosa==0.10.2.post1
moviepy==1.0.3
numpy==2.1.3